+++ package.json
@@ -44,6 +44,7 @@
   },
   "devDependencies": {
     "autoprefixer": "^7.2.5",
+    "axios": "^0.17.1",
     "babel": "^6.23.0",
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.2",
