+++ public/sass/partials/_core.sass
@@ -56,5 +56,3 @@ body
   text-align: center
   width: 90%
   max-width: 480px
-
-/* ----------- Replace: For links with no text -----------*/
