+++ public/sass/styles.sass
@@ -1,9 +1,6 @@
 /* Main Styles: Import all the things here                                  */
 /*----------------------------------------------------------------------------------*/
 
-/* ----------- Google Font -----------*/
-@import url('https://fonts.googleapis.com/css?family=Fira+Sans:200,300,300i,400,400i,600,600i,700,700i|Pacifico')
-
 /* ----------- Normalize -----------*/
 @import '~normalize.css'
